Split templates and classes
