+++ src/components/film-card/film-card.js
@@ -1,0 +1,25 @@
+import { createElement } from '../../utils';
+import { createFilmCardTemplate } from './film-card-tpl';
+
+export default class FilmsSectionComponent {
+  constructor(card) {
+    this._card = card;
+    this._element = null;
+  }
+
+  getTemplate() {
+    return createFilmCardTemplate(this._card);
+  }
+
+  getElement() {
+    if (!this._element) {
+      this._element = createElement(this.getTemplate());
+    }
+
+    return this._element;
+  }
+
+  removeElement() {
+    this._element = null;
+  }
+}
