+++ src/components/film-section/film-section.js
@@ -1,0 +1,24 @@
+import { createElement } from '../../utils';
+import { createFilmsSectionTemplate } from './film-section-tpl';
+
+export default class FilmCardComponent {
+  constructor() {
+    this._element = null;
+  }
+
+  getTemplate() {
+    return createFilmsSectionTemplate();
+  }
+
+  getElement() {
+    if (!this._element) {
+      this._element = createElement(this.getTemplate());
+    }
+
+    return this._element;
+  }
+
+  removeElement() {
+    this._element = null;
+  }
+}
