+++ src/components/film-statistics/film-statistics-tpl.js
@@ -1,0 +1,5 @@
+export const createFilmsStatistics = (movies) => {
+    return `<section class="footer__statistics">
+              <p>${movies.length} movies inside</p>
+            </section>`;
+}
