+++ src/components/navigation/navigation.js
@@ -1,0 +1,25 @@
+import { createElement } from '../../utils';
+import { createNavigationTemplate } from './navigation-tpl';
+
+export default class NavigationComponent {
+  constructor(filters) {
+    this._filters = filters;
+    this._element = null;
+  }
+
+  getTemplate() {
+    return createNavigationTemplate(this._filters);
+  }
+
+  getElement() {
+    if (!this._element) {
+      this._element = createElement(this.getTemplate());
+    }
+
+    return this._element;
+  }
+
+  removeElement() {
+    this._element = null;
+  }
+}
