+++ src/components/no-films/no-films-tpl.js
@@ -1,0 +1,3 @@
+export const createNoFilmsTemplate = () => {
+    return `<h2 class="films-list__title">There are no movies in our database</h2>`;
+}
