+++ src/components/no-films/no-films.js
@@ -1,0 +1,24 @@
+import { createElement } from '../../utils';
+import { createNoFilmsTemplate } from './no-films-tpl';
+
+export default class NoFilmsComponent {
+  constructor() {
+    this._element = null;
+  }
+
+  getTemplate() {
+    return createNoFilmsTemplate();
+  }
+
+  getElement() {
+    if (!this._element) {
+      this._element = createElement(this.getTemplate());
+    }
+
+    return this._element;
+  }
+
+  removeElement() {
+    this._element = null;
+  }
+}
