+++ src/utils.js
@@ -5,7 +5,7 @@ export const createElement = (template) => {
   newElement.innerHTML = template;
 
   return newElement.firstChild;
-};
+}
 
 export const render = (container, element, place) => {
   switch (place) {
@@ -17,60 +17,39 @@ export const render = (container, element, place) => {
       container.append(element);
       break;
   }
-};
+}
 
-export const getRandomWords = (data) => data[Math.floor(Math.random() * data.length)];
-
-export const getRandomIntegerNumber = (min, max) => Math.floor(Math.random() * (max - min + 1)) + min;
-
-export const getRandomFloatNumber = (min, max) => Math.floor((Math.random() * ((max - min)) + 1) * 10) / 10;
+export const getRandomWords = (data) => data[Math.floor(Math.random() * data.length)]
+export const getRandomIntegerNumber = (min, max) => Math.floor(Math.random() * (max - min + 1)) + min
+export const getRandomFloatNumber = (min, max) => Math.floor((Math.random() * ((max - min)) + 1) * 10) / 10
+export const getRandomArray = (array) => array.sort(() => Math.random() - 0.5)
+export const getRandomGenres = (array) => getRandomArray(array).slice(0, 3)
+export const getTextSentenceCase = (sentence) => `${sentence[0].toUpperCase()}${sentence.slice(1)}`.trim()
+export const getRandomBoolean = () => Math.random() > 0.5
 
 export const getRandomArrayItem = (array) => {
   const randomIndex = getRandomIntegerNumber(0, array.length - 1);
 
   return array[randomIndex];
-};
-
-export const getRandomArray = (array) => {
-
-  return array.sort(() => Math.random() - 0.5);
-};
-
-export const getRandomGenres = (array) => {
-
-  return getRandomArray(array).slice(0, 3);
-};
-
-export const getTextSentenceCase = (sentence) => `${sentence[0].toUpperCase()}${sentence.slice(1)}`.trim();
-
-export const getRandomBoolean = () => Math.random() > 0.5;
+}
 
 export const generateText = (min, max, array) => {
   let string = ``;
-  new Array(getRandomIntegerNumber(min, max))
-    .fill(``)
-    .map(() => {
-      string += `${getRandomWords(array)} `;
-    });
+  [...Array(getRandomIntegerNumber(min, max))].map(() => string += `${getRandomWords(array)} `)
 
   return getTextSentenceCase(string.substr(Range.MIN_STRING, Range.MAX_STRING));
-};
+}
 
-export const generateComments = (min, max, array) => {
-  return new Array(getRandomIntegerNumber(min, max))
-    .fill(``)
-    .map(() => generateText(min, max, array));
-};
+export const generateComments = (min, max, array) => [...Array(getRandomIntegerNumber(min, max))].map(() => generateText(min, max, array))
 
 export const getRandomDate = () => {
   const targetDate = new Date();
   const sign = Math.random() > 0.5 ? 1 : -1;
   const diffValue = sign * getRandomIntegerNumber(0, 7);
-
   targetDate.setDate(targetDate.getDate() + diffValue);
 
   return targetDate;
-};
+}
 
 export const getFormatDate = (date) => {
   let dd = date.getDate();
@@ -85,6 +64,6 @@ export const getFormatDate = (date) => {
   }
 
   return `${dd} ${mm} ${yy}`;
-};
+}
 
-export const isEscPressed = (evt) => evt.key === `Escape` || evt.key === `Esc`;
+export const isEscPressed = (evt) => evt.key === `Escape` || evt.key === `Esc`
